Configure a default datasource directly from ENV VARS or config (#27)
Add new config values to configure a fallback datasource in elasticsearch. When returning the complete list of datasources we are using the local URL to do a proxy call. It's when we are doing the proxy call that we ask for a specific datasource and we load the elasticsearch URL.
Fix default home dashboard to render CPU data correctly (and a typo).